update mybatis demo

+++ docs/examples/JDBC/mybatisplus-demo/src/test/java/com/taosdata/example/mybatisplusdemo/mapper/TemperatureMapperTest.java
@@ -107,7 +107,7 @@ public class TemperatureMapperTest {
      * **/
     @Test
     public void testSelectCount() {
-        int count = mapper.selectCount(null);
+        long count = mapper.selectCount(null);
         Assert.assertEquals(10, count);
     }
